#162: Moving UIState from debuda.py to dbd/tt-uistate.py
