+++ dbd/debuda_commands/core-debug-regs.py
@@ -18,7 +18,7 @@ Examples:
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_util as util
 from dbd.tt_coordinate import OnChipCoordinate
