+++ dbd/debuda_commands/dump-debug-buffer.py
@@ -22,7 +22,7 @@ Examples:
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_device
 from dbd import tt_util as util
