+++ dbd/debuda_commands/dump-gpr.py
@@ -28,7 +28,7 @@ command_metadata = {
 
 import tabulate
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_commands
 from dbd import tt_util as util
