+++ dbd/debuda_commands/gdb.py
@@ -15,7 +15,7 @@ Examples:
 """
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_util as util
