+++ dbd/debuda_commands/jtag-read-axi.py
@@ -28,7 +28,7 @@ command_metadata = {
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_device
 from dbd.tt_coordinate import OnChipCoordinate
