+++ dbd/debuda_commands/pci-write-xy.py
@@ -26,7 +26,7 @@ command_metadata = {
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_device
 from dbd.tt_coordinate import OnChipCoordinate
