+++ dbd/debuda_commands/print-stream.py
@@ -26,7 +26,7 @@ command_metadata = {
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_stream, tt_util as util
 from dbd.tt_coordinate import OnChipCoordinate
