+++ dbd/debuda_commands/tile-summary.py
@@ -18,7 +18,7 @@ Examples:
 
 from docopt import docopt
 
-from debuda import UIState
+from dbd.tt_uistate import UIState
 
 from dbd import tt_util as util
 from dbd.tt_debuda_ifc import debuda_server_not_supported
